pyproject.toml
[project]
name = "hmpps-person-match"
version = "0.1.0"
description = ""
authors = []

requires-python = ">=3.13"
dependencies = [
  "splink==4.0.11",
  "duckdb==1.3.2",
  "psycopg2-binary==2.9.11",
  "pytz==2025.2",
  "azure-monitor-opentelemetry==1.8.1",
  "fastapi==0.122.0",
  "uvicorn==0.38.0",
  "authlib==1.6.5",
  "cachetools==6.2.2",
  "pyjwt==2.10.1",
  "cryptography==46.0.3",
  "alembic==1.17.2",
  "asyncpg==0.30.0",
  "greenlet==3.2.4",
  "opentelemetry-instrumentation-sqlalchemy==0.57b0",
  # h11 is required by uvicorn and httpx to fix CVE-2025-43859
  "h11==0.16.0",
  "types-authlib>=1.6.5.20251005",
  "types-cachetools>=6.2.0.20251022",
  "types-requests>=2.32.4.20250913",
]

[tool.uv]
package = true
default-groups = ["dev"]

[dependency-groups]
dev = [
  "ruff==0.14.6",
  "pyyaml==6.0.3",
  "pytest==9.0.1",
  "pytest-cov==7.0.0",
  "pytest-asyncio==1.3.0",
  "faker==38.2.0",
  "httpx==0.28.1",
  "requests-mock==1.12.1",
  "mypy>=1.18.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
dummy-variable-rgx = "^_$"
select = [
    # pycodestyle
    "E",
    "W",
    # pep8-naming
    "N",
    # pyupgrade
    "UP",
    # Pyflakes
    "F",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-2020
    "YTT",
    # flake8-blind-except
    "BLE",
    # flake8-bandit
    "S",
    # flake8-quotes
    "Q",
    # flake8-logging
    "LOG",
    # flake8-logging-format
    "G",
    # flake8-no-pep420
    "INP",
    # flake8-pie
    "PIE",
    # flake8-self
    "SLF",
    # isort
    "I",
    # pylint
    "PL",
    # fastapi
    "FAST",
    # flake8-async
    "ASYNC",
    # flake8-datetimez
    "DTZ",
    # flake8-debugger
    "T10",
    # flake8-tidy-imports
    "TID",
    # ERA
    "ERA",
    # refurb
    "FURB",
    # ruff-specific-rules
    "RUF",
    # perflint
    "PERF",
    # flake8-annotation
    "ANN",
]
ignore = [
    # magic-value-comparison
    "PLR2004"
]

[tool.ruff.lint.extend-per-file-ignores]
# Also ignore `S: flake8-bandit` in all `test_*.py` files.
"test_*.py" = ["S"]
# ignores name convention as overriding package function
"log_formatter.py" = ["N802"]
"hmpps_cpr_splink/*" = [
    # flake8-self
    # keep this as we need to access Splink internals - not a problem as we own it
    "SLF",
]
"hmpps_cpr_splink/tests/*" = [
    # flake8-bandit
    "S",
    # flake8-no-pep420
    "INP",
]
